Fix cycle dependency
